Separate eventListeners into a separate module
